% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-active_alerts.R
\name{alerts}
\alias{alerts}
\title{Get active NWS alerts}
\usage{
alerts()
}
\value{
A dataframe containing various columns identifying and describing alerts
}
\description{
Get active NWS alerts
}
\examples{
# Get all the red flag warnings
red_flags <- alerts()
red_flags <- filter(red_flags, grepl("red flag", red_flags$headline, ignore.case = TRUE))
}
