% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-auxiliary.R
\name{dir_as_char}
\alias{dir_as_char}
\title{Convert wind direction from a numeric to a character}
\usage{
dir_as_char(direction)
}
\arguments{
\item{direction}{A string corresponding to direction ("N", "NNE", etc.)}
}
\value{
Returns a character (N, NNE, NE, etc.)
}
\description{
Convert wind direction from a numeric to a character
}
\examples{
dir_as_char(330)
}
