% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-auxiliary.R
\name{dir_as_integer}
\alias{dir_as_integer}
\title{Convert wind direction from a character to an integer}
\usage{
dir_as_integer(direction)
}
\arguments{
\item{direction}{The direction as a string (N, NNE, NE, ENE, E, ESE, etc.)}
}
\value{
An integer representing degrees clockwise from north
}
\description{
Convert wind direction from a character to an integer
}
\examples{
dir_as_integer("NNW")
}
