% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-point_forecast.R
\name{..point_forecast}
\alias{..point_forecast}
\title{Get the point forecast AND the timezone for the point, all in one function call.}
\usage{
..point_forecast(lat, lon, timezone = -1, dir_numeric = FALSE)
}
\arguments{
\item{lat}{Latitude}

\item{lon}{Longitude}

\item{timezone}{Timezone, -1 for local time}

\item{dir_numeric}{Logical; TRUE for numeric directions, FALSE for character directions}
}
\value{
A list containing point forecast sf and the timezone
}
\description{
Get the point forecast AND the timezone for the point, all in one function call.
}
\examples{
\donttest{
..point_forecast(33, -80)
}
}
