% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-point_data.R
\name{.point_data}
\alias{.point_data}
\title{Get the JSON Data for a Point}
\usage{
.point_data(lat, lon)
}
\arguments{
\item{lat}{Latitude of the point to fetch data for.}

\item{lon}{Longitude of the point to fetch data for.}
}
\value{
A nested list containing NWS point data.
}
\description{
Get the JSON Data for a Point
}
\examples{
\donttest{
.point_data(lat = 40.71427000, lon = -74.00597000) \%>\% data.frame()
}
}
