% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dublinairport}
\alias{dublinairport}
\title{Dublin Airport Data Constructor}
\usage{
dublinairport(data)
}
\arguments{
\item{data}{A data frame to which the \code{dublinairport} class will be assigned.}
}
\value{
A data frame with the class \code{dublinairport}.
}
\description{
A constructor function for assigning the class \code{dublinairport} to a data frame.
This function checks that the input is a data frame and adds a custom class for easier identification and use within the package.
}
\examples{
 #Example 1 : assigning class to a data set
data <- data.frame(Month = c("January", "February"),
                   `Mean Temperature` = c(5.0, 5.5),
                   `Mean Daily Sunshine (hours)` = c(2.5, 2.7),
                   `Mean Rainfall` = c(75.0, 62.0))
dublin_airport_obj <- dublinairport(data)
class(dublin_airport_obj)

#Example 2: Checking for invalid input to the function
#This will give an error because the input needs to be a data frame
invalid_input <- list(a=1,b=2)
try(dublinairport(invalid_input))
}
