% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get-Weather.R
\name{get_daily}
\alias{get_daily}
\alias{get_hourly}
\title{Get Weather Data}
\usage{
get_daily(stationid)

get_hourly(stationid, year)
}
\arguments{
\item{stationid}{USAF-WBAN station ID}

\item{year}{Year}
}
\description{
Get Weather Data
}
\section{Functions}{
\itemize{
\item \code{get_hourly}: 
}}

