% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather.R
\name{get_forecast_endpoints}
\alias{get_forecast_endpoints}
\title{Get endpoints for forecasts given lat and lon}
\usage{
get_forecast_endpoints(lat, lon)
}
\arguments{
\item{lat}{latitude of location}

\item{lon}{longitude of location}
}
\description{
Get endpoints for forecasts given lat and lon
}
