% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather.R
\name{get_forecast_properties}
\alias{get_forecast_properties}
\title{Get properties from list of all data}
\usage{
get_forecast_properties(lat, lon)
}
\arguments{
\item{lat}{latitude}

\item{lon}{longitude}
}
\description{
Get properties from list of all data
}
