% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather.R
\name{get_grid_data}
\alias{get_grid_data}
\title{Get hourly forecast}
\usage{
get_grid_data(lat, lon)
}
\arguments{
\item{lat}{latitude of location}

\item{lon}{longitude of location}
}
\description{
Get hourly forecast
}
