% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_hourly_data.R
\name{get_hourly_table}
\alias{get_hourly_table}
\title{Process the hourly table into a nicer format}
\usage{
get_hourly_table(hourly_data, hours = 12)
}
\arguments{
\item{hourly_data}{data returned by the hourly forecast API}
}
\description{
Process the hourly table into a nicer format
}
