% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_location.R
\name{get_location}
\alias{get_location}
\title{Get geocoded lat/lon from the census}
\usage{
get_location(address)
}
\arguments{
\item{address}{address to get the lat and lon from}
}
\description{
Get geocoded lat/lon from the census
}
