% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dublinairport.R
\name{load_dublinairport}
\alias{load_dublinairport}
\title{Load the Dublin Airport Dataset}
\usage{
load_dublinairport()
}
\value{
A data frame containing the metric averages for Dublin Airport, with the following columns.
\describe{
\item{\code{Mean Temperature}}{The mean temperature in degrees celcius}
\item{\code{Mean Daily Sunshine (hours)}}{The mean daily sunshine duration in hours.}
\item{\code{Mean Rainfall}}{The total monthly rainfall in millimeters.}
}
}
\description{
This function loads the Dublin Airport weathR package.
The dataset contains climate averages recorded at Dublin airport for the period of 1981 - 2010.
}
\details{
IF the dataset is not already loaded in the environment, the function automatically loads it.
It also varifies the structure of the dataset before returning it.
}
\examples{
#Load the dataset
data <- load_dublinairport()

#View the first few rows
head(data)

}
