% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\docType{data}
\name{nightfall_styles}
\alias{nightfall_styles}
\title{ANSI styles based on Paul Tol's color notes}
\format{
An object of class \code{list} of length 17.
}
\usage{
nightfall_styles
}
\description{
ANSI styles based on Paul Tol's color notes
}
\keyword{datasets}
