% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_hourly_data.R
\name{parse_time}
\alias{parse_time}
\title{Parse the weather time period format that's ugly}
\usage{
parse_time(weather_times)
}
\arguments{
\item{weather_times}{start or end times for interval
as returned by the API}
}
\description{
Parse the weather time period format that's ugly
}
