% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_hourly_data.R
\name{plot_hourly_temperature}
\alias{plot_hourly_temperature}
\title{Plot hourly data in various ways in the console}
\usage{
plot_hourly_temperature(hourly_data, hours = 12)
}
\description{
Plot hourly data in various ways in the console
}
