% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_hourly_data.R
\name{plot_rain_chance}
\alias{plot_rain_chance}
\title{Plot hourly data in various ways in the console}
\usage{
plot_rain_chance(hourly_data, hours = 12)
}
\description{
Plot hourly data in various ways in the console
}
