% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dublinairport.R
\name{plot_summarise}
\alias{plot_summarise}
\title{Plot and Animate Dublin Airport Climate Trends}
\usage{
plot_summarise(data)
}
\description{
Creates an animated plot visualising climate data from Dublin Airport, including trends in mean temperature,
mean daily sunshine, and mean rainfall over the months of the year.
}
