% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-point_data.R
\name{point_data}
\alias{point_data}
\title{Get NWS Metadata for a Point}
\usage{
point_data(lat, lon)
}
\arguments{
\item{lat}{Latitude.}

\item{lon}{Longitude.}
}
\value{
A simple features point object with NWS metadata.
}
\description{
Get NWS Metadata for a Point
}
\examples{
\donttest{
point_data(lat = 40.71427000, lon = -74.00597000)
}
}
