% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-point_forecast.R
\name{point_forecast}
\alias{point_forecast}
\title{Point Forecast Data}
\usage{
point_forecast(lat, lon, timezone = -1, dir_numeric = FALSE)
}
\arguments{
\item{lat}{Latitude.}

\item{lon}{Longitude.}

\item{timezone}{The nominal timezone for the forecast. One of \code{OlsonNames()} or \code{-1} for local time. Defaults to \code{-1}.}

\item{dir_numeric}{\code{TRUE} for numeric directions, \code{FALSE} for character directions; defaults to \code{FALSE}.}
}
\value{
Simple features object with forecast meteorological values.
}
\description{
Point Forecast Data
}
\examples{
\donttest{
point_forecast(lat = 40.71427000, lon = -74.00597000, dir_numeric = TRUE)
}
}
