% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-point_data.R
\name{point_station}
\alias{point_station}
\title{Find Nearest ASOS/AWOS Station}
\usage{
point_station(lat, lon)
}
\arguments{
\item{lat}{Latitude.}

\item{lon}{Longitude.}
}
\value{
A string corresponding to an ASOS or AWOS station.
}
\description{
Find Nearest ASOS/AWOS Station
}
\examples{
# Gets the observation data as an sf associated with a point
\donttest{
point_station(lat = 42, lon = -80) \%>\% station_obs() \%>\% data.frame()
}
}
