% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-point_forecast.R
\name{point_tomorrow}
\alias{point_tomorrow}
\title{Fetch the weather forecast for a lat/lon point tomorrow, denominated in the given timezone.}
\usage{
point_tomorrow(lat, lon, timezone = -1, dir_numeric = FALSE, short = TRUE)
}
\arguments{
\item{lat}{Latitude}

\item{lon}{Longitude}

\item{timezone}{A timezone to denominate timestamps in (must be one of the Olson Names)}

\item{dir_numeric}{TRUE/FALSE; Should the wind direction be numeric, or character?}

\item{short}{TRUE/FALSE; Should the forecast be for only tomorrow, or both the remainder of today and tomorrow?}
}
\value{
Returns an sf with forecast values
}
\description{
Fetch the weather forecast for a lat/lon point tomorrow, denominated in the given timezone.
}
\examples{
# Get the maximum temperature in tomorrows forecast
\donttest{
point_tomorrow(lat = 33, lon = -80)
}
}
