% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather.R
\name{properties}
\alias{properties}
\title{get properties from weather json}
\usage{
properties(weather_json)
}
\arguments{
\item{weather_json}{json from get_request}
}
\description{
get properties from weather json
}
