% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-station_data.R
\name{station_coords}
\alias{station_coords}
\title{Station Coordinates}
\usage{
station_coords(station_id)
}
\arguments{
\item{station_id}{The station identifier (ex: KDEN, KBOS, KNYC, etc).}
}
\value{
Named list with latitude and longitude like: \code{c("lat" = x, "lon" = y)}.
}
\description{
Station Coordinates
}
\examples{
\dontrun{
station_coords("KBOS") \%>\%
  data.frame(lat = .[["lat"]], lon = .[["lon"]]) \%>\%
  select(lat, lon) \%>\%
  st_as_sf(coords = c("lon", "lat"), crs = 4326)
}
}
