% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-station_forecast.R
\name{station_forecast}
\alias{station_forecast}
\title{Fetch the NWS hourly forecast at a given AWOS or ASOS station identifier.}
\usage{
station_forecast(station_id, timezone = -1, dir_numeric = FALSE)
}
\arguments{
\item{station_id}{The station identifier for which to get the observations}

\item{timezone}{The timezone to denominate timestamps in. Defaults to local time.}

\item{dir_numeric}{Should the wind direction be returned as a numeric, or a character? TRUE uses numeric wind directions, FALSE uses characters.}
}
\value{
An sf with coordinates for the AWOS/ASOS station, time, and forecast values.
}
\description{
Fetch the NWS hourly forecast at a given AWOS or ASOS station identifier.
}
\examples{
\donttest{
station_forecast("KBOS") \%>\% data.frame() \%>\% dplyr::select(-geometry)
}
}
