% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-station_observations.R
\name{station_obs}
\alias{station_obs}
\title{Fetch station observations for an ASOS or AWOS ID}
\usage{
station_obs(station_id, timezone = -1, dir_numeric = FALSE)
}
\arguments{
\item{station_id}{The AWOS or ASOS station ID}

\item{timezone}{The timezone to denominate timestamps in. Defaults to local time}

\item{dir_numeric}{Should the wind direction be returned as a numeric, or a character? TRUE uses numeric wind directions, FALSE uses characters}
}
\value{
An sf with coordinates for the AWOS/ASOS station, time, and observed values.
}
\description{
Fetch station observations for an ASOS or AWOS ID
}
\examples{
# Gets the highest observed temperatures in Boston
\donttest{
station_obs("KBOS") \%>\% data.frame() \%>\%
  dplyr::select(-geometry) \%>\%
  dplyr::filter(temp == max(.$temp))
}
}
