% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-station_data.R
\name{station_point}
\alias{station_point}
\title{Fetch station coordinates as a point}
\usage{
station_point(station_id)
}
\arguments{
\item{station_id}{The ASOS or AWOS station ID (ex: KDEN, KBOS, KNYC, etc)}
}
\value{
an sf point
}
\description{
Fetch station coordinates as a point
}
\examples{
# plots the ASOS station for KDEN using tmap
\dontrun{
tmap_mode("view")
station_point("KDEN") \%>\% tm_shape() + tm_dots(size = .1)
}
}
