% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-station_today.R
\name{station_today}
\alias{station_today}
\title{Get todays weather data, given a station ID}
\usage{
station_today(station_id, timezone = -1, dir_numeric = FALSE)
}
\arguments{
\item{station_id}{The ASOS or AWOS station ID (Ex: KDEN, KNYC, etc.)}

\item{timezone}{The timezone to denominate results in. Must be one of OlsonNames().}

\item{dir_numeric}{Logical; Should the wind direction be returned as a numeric, or a character?}
}
\value{
A sf object with timestamps and weather values
}
\description{
Get todays weather data, given a station ID
}
\examples{
station_today("KNYC")
}
