% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-station_forecast.R
\name{station_tomorrow}
\alias{station_tomorrow}
\title{Get tomorrows forecast for a given station_id}
\usage{
station_tomorrow(station_id, timezone = -1, dir_numeric = FALSE, short = TRUE)
}
\arguments{
\item{station_id}{Station identifier (ex: KDEN, KBOS, KNYC, etc).}

\item{timezone}{The nominal timezone for the forecast. One of \code{OlsonNames()} or \code{-1} for local time. Defaults to \code{-1}.}

\item{dir_numeric}{\code{TRUE} for numeric directions, \code{FALSE} for character directions; defaults to \code{FALSE}.}

\item{short}{\code{TRUE} for only tomorrow, \code{FALSE} for today and tomorrow; defaults to \code{TRUE}.}
}
\value{
Simple features point corresponding to the given station identifier with recent meteorological forecast values.
}
\description{
Get tomorrows forecast for a given station_id
}
\examples{
station_tomorrow("KBOS")

}
