% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-station_data.R
\name{station_tz}
\alias{station_tz}
\title{Fetch station timezone}
\usage{
station_tz(station_id)
}
\arguments{
\item{station_id}{The ASOS or AWOS station ID (ex: KDEN, KBOS, KNYC, etc)}
}
\value{
Returns a string corresponding to a timezone from OlsonNames()
}
\description{
Fetch station timezone
}
\examples{
# Converts the current time into the timezone for the Denver International Airport ASOS station.
donttest{
Sys.time() \%>\% lubridate::force_tz(tzone = station_tz("KDEN"))
}
}
