% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-point_data.R
\name{stations_near}
\alias{stations_near}
\title{Fetch stations near a lat/lon point, and within that lat/lon's forecast zone}
\usage{
stations_near(lat, lon)
}
\arguments{
\item{lat}{Latitude for which to find nearby stations}

\item{lon}{Longitude for which to find nearby stations}
}
\value{
An sf object with station identifiers, geometry as coordinates, and their euclidian distance (in miles) to the station provided.
}
\description{
Fetch stations near a lat/lon point, and within that lat/lon's forecast zone
}
\examples{
# Plot the a station with given points and the nearby stations in a tmap
\donttest{
stations_near(lat = 33, lon = -80) \%>\%
  tm_shape() +
  tm_dots(size = .1) +
   tm_shape(st_point(c(-80, 33)) \%>\% st_sfc(crs = 4326)) + tm_dots(size = .1, col = "red3")
}
}
