% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.dublinairport.R
\name{summarise.dublinairport}
\alias{summarise.dublinairport}
\title{Summarise Dublin Airport Weather Data}
\usage{
\method{summarise}{dublinairport}(
  object,
  filter_var = NULL,
  filter_value = NULL,
  summary_func = mean,
  correlation = FALSE,
  ...
)
}
\arguments{
\item{object}{A dataset of the class "dublin airport".}

\item{filter_var}{A character string specifying the name of the column to filter by.
If 'NULL', no filtering occurs. The default is \code{NULL}.}

\item{filter_value}{A numeric value that filters rows where \code{filter_var} is greater
than or equal to this value. The default is \code{NULL}.}

\item{summary_func}{A function to calculate summary statistics for the numeric columns.
The default is \code{mean}.}

\item{correlation}{A logical indicating whether to include the correlation matrix in the output. Defaults to \code{TRUE}.}

\item{...}{Additional arguments used by \code{summary_func}.}
}
\value{
A list containing:
 \item {summary}{A tibble of the summary statistic for the numeric columns.}
 \item {correlation_matrix}{A correlation matrix of the numeric columns (if 'correlation = TRUE').}
}
\description{
Summarises numeric columns in the `dublinairport` dataset. Allows optional filtering
of rows based on a column and value. Calculates a correlation matrix if specified.
}
\examples{
data(dublin_airport)

result <- summarise.dublinairport(dublin_airport)

result_with_correlation <- summarise.dublinairport(dublin_airport, correlation = TRUE)

summarise.dublinairport(dublin_airport, filter_var = "Mean Rainfall", filter_value = 60)

summarise.dublinairport(dublin_airport, filter_var = "Mean Temperature", filter_value = 10)

result_quantiles <- summarise.dublinairport(dublin_airport,
                                            summary_func = quantile,
                                            probs = c(0.25, 0.5, 0.75))
}
