% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{temperature_colors}
\alias{temperature_colors}
\title{Get color palette for temperature colors}
\usage{
temperature_colors(temps, styles)
}
\arguments{
\item{temp}{temperature vector}

\item{palette}{function returning color palette}
}
\description{
Get color palette for temperature colors
}
