% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-active_alerts.R
\name{alerts}
\alias{alerts}
\title{National Weather Service Alerts}
\usage{
alerts()
}
\value{
Dataframe containing various columns identifying and describing alerts.
}
\description{
National Weather Service Alerts
}
\examples{
\donttest{
alerts()
}
}
