% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-auxiliary.R
\name{dir_as_char}
\alias{dir_as_char}
\title{Convert Wind Direction from Numeric to Character}
\usage{
dir_as_char(direction)
}
\arguments{
\item{direction}{Numeric degrees clockwise from north, (0 = N, 45 = NE, etc).}
}
\value{
Character direction (N, NNE, NE, etc).
}
\description{
Convert Wind Direction from Numeric to Character
}
\examples{
dir_as_char(330)
}
