% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-auxiliary.R
\name{dir_as_integer}
\alias{dir_as_integer}
\title{Convert Wind Direction from a Character to an Integer}
\usage{
dir_as_integer(direction)
}
\arguments{
\item{direction}{A direction, as a string (N, NNE, NE, ENE, E, ESE, etc).}
}
\value{
An integer representing degrees clockwise from north.
}
\description{
Convert Wind Direction from a Character to an Integer
}
\examples{
dir_as_integer("NNW")
}
