% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-point_data.R
\name{.point_data}
\alias{.point_data}
\title{Get the unprocessed http response for point data corresponding to a given latitude/longitude from the NWS}
\usage{
.point_data(lat, lon)
}
\arguments{
\item{lat}{Latitude of the point to fetch}

\item{lon}{Longitude of the point to fetch}
}
\value{
A nested list containing NWS point data
}
\description{
Get the unprocessed http response for point data corresponding to a given latitude/longitude from the NWS
}
\examples{
.point_data(lat = 40.71427000, lon = -74.00597000) \%>\% data.frame()
}
