% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-point_forecast.R
\name{.point_forecast}
\alias{.point_forecast}
\title{Get point forecast data as a raw nested json list}
\usage{
.point_forecast(lat, lon)
}
\arguments{
\item{lat}{Latitude}

\item{lon}{Longitude}
}
\value{
Returns the json data as a nested list
}
\description{
Get point forecast data as a raw nested json list
}
\examples{
.point_forecast(33, -80)
}
