% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-station_data.R
\name{.station_data}
\alias{.station_data}
\title{Fetch station metadata data as a nested list}
\usage{
.station_data(station_id)
}
\arguments{
\item{station_id}{The ASOS or AWOS station ID (ex: KDEN, KBOS, KNYC, etc)}
}
\value{
Returns all station data provided for the NWS
}
\description{
Fetch station metadata data as a nested list
}
\examples{
\donttest{
.station_data("KDEN")
}
}
