% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dublinairport.R
\name{plot_summarise}
\alias{plot_summarise}
\title{Plot Climate Data for Dublin Airport}
\usage{
plot_summarise(data)
}
\arguments{
\item{data}{A data frame containing climate data with specific column requirements.}
}
\value{
A gganimate animation object visualising climate trends
}
\description{
Creates an animated line plot visualising climate trends including temperature,
daily sunshine, and rainfall for Dublin Airport across months of the year.
}
\details{
The function requires the following columns:
- 'Month'
- 'Mean Temperature'
- 'Mean Daily Sunshine (hours)'
- 'Mean Rainfall'

The function performs the following transformations:
- Converts month to a factor with a specific order
- Pivots data to long format for plotting
- Scales rainfall data by dividing by 10
- Creates an animated line plot with revealing transition
}
\examples{
\dontrun{
anim <- plot_summarise(dublin_airport)
print(anim)
}

}
\author{
Sean Carroll - <\email{sean.carroll.2021@mu.ie.ie}>
}
