% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-point_data.R
\name{point_data}
\alias{point_data}
\title{Gets a subset of NWS point metadata for a given lat/lon}
\usage{
point_data(lat, lon)
}
\arguments{
\item{lat}{Latitude of the point to fetch}

\item{lon}{Longitude of the point to fetch}
}
\value{
A sf point object with NWS metadata
}
\description{
Gets a subset of NWS point metadata for a given lat/lon
}
\examples{
point_data(lat = 40.71427000, lon = -74.00597000)
}
