% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-point_today.R
\name{point_today}
\alias{point_today}
\title{Get the weather observed thus far today at the nearest ASOS station to a point which is in the same forecast zone as the given point.}
\usage{
point_today(lat, lon, timezone = -1, dir_numeric = FALSE)
}
\arguments{
\item{lat}{Latitude of a point}

\item{lon}{Longitude of a point}

\item{timezone}{Timezone to denominate results in; -1 for local time}

\item{dir_numeric}{Logical; TRUE for numeric wind directions, FALSE for character wind directions}
}
\value{
A sf object with weather observations for today
}
\description{
Get the weather observed thus far today at the nearest ASOS station to a point which is in the same forecast zone as the given point.
}
\examples{
# Produce a GT summary of the weather thus far for a given lat/lon
\donttest{
install.packages("gt")
library(gt)
install.packages("snakecase")
library(snakecase)

point_today(lat = 33, lon = -80) \%>\%
as.data.frame() \%>\%
 dplyr::mutate(time = as.POSIXct(time) \%>\% format("\%H:\%M")) \%>\%
  dplyr::select(time, temp, dewpoint, humidity, wind_speed) \%>\%
  gt::gt() \%>\%
  gt::tab_header(title = "Weather at latitude 33, longitude -80 for today") \%>\%
  gt::tab_style(
    style = gt::cell_borders(
      sides = "right",
      color = "lightgray",
      style = "dashed",
      weight = px(1)
      ),
      locations = cells_body(columns = everything())
    ) \%>\%
   gt::tab_style(
    style = gt::cell_text(align = "center"),
     locations = gt::cells_column_labels(columns = everything())
   ) \%>\%
   gt::tab_style(
    style = gt::cell_borders(sides = "right", color = "lightgray", weight = px(2)),
    locations = gt::cells_body(columns = 1)
    ) \%>\%
    gt::cols_label_with(fn = ~ snakecase::to_title_case(.x))
 }
}
