% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-point_forecast.R
\name{point_tomorrow}
\alias{point_tomorrow}
\title{Weather Forecast for Tomorrow at a Point}
\usage{
point_tomorrow(lat, lon, timezone = -1, dir_numeric = FALSE, short = TRUE)
}
\arguments{
\item{lat}{Latitude.}

\item{lon}{Longitude.}

\item{timezone}{The nominal timezone for the forecast. One of \code{OlsonNames()} or \code{-1} for local time. Defaults to \code{-1}.}

\item{dir_numeric}{\code{TRUE} for numeric directions, \code{FALSE} for character directions; defaults to \code{FALSE}.}

\item{short}{\code{TRUE} for only tomorrow, \code{FALSE} for today and tomorrow; defaults to \code{TRUE}.}
}
\value{
Simple features object with forecast meteorological values.
}
\description{
Weather Forecast for Tomorrow at a Point
}
\examples{
\donttest{
point_tomorrow(lat = 33, lon = -80)
}
}
