% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-station_data.R
\name{station_coords}
\alias{station_coords}
\title{Fetch station coordinates}
\usage{
station_coords(station_id)
}
\arguments{
\item{station_id}{The ASOS or AWOS station ID (ex: KDEN, KBOS, KNYC, etc)}
}
\value{
Returns a named list with latitude and longitude like c("lat" = x, "lon" = y)
}
\description{
Fetch station coordinates
}
\examples{
# Converts station data to a sf point
\dontrun{
station_coords("KBOS") \%>\%
  data.frame(lat = .[["lat"]], lon = .[["lon"]]) \%>\%
  select(lat, lon) \%>\%
  st_as_sf(coords = c("lon", "lat"), crs = 4326)
}
}
