% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-station_forecast.R
\name{station_forecast}
\alias{station_forecast}
\title{Station Forecast}
\usage{
station_forecast(station_id, timezone = -1, dir_numeric = FALSE)
}
\arguments{
\item{station_id}{Station identifier (ex: KDEN, KBOS, KNYC, etc).}

\item{timezone}{The nominal timezone for the forecast. One of \code{OlsonNames()} or \code{-1} for local time. Defaults to \code{-1}.}

\item{dir_numeric}{\code{TRUE} for numeric directions, \code{FALSE} for character directions; defaults to \code{FALSE}.}
}
\value{
Simple features point corresponding to the given station identifier with recent meteorological forecast values.
}
\description{
Station Forecast
}
\examples{
\donttest{
station_forecast("KBOS") \%>\% data.frame() \%>\% dplyr::select(-geometry)
}
}
