% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-station_today.R
\name{station_today}
\alias{station_today}
\title{Get todays weather data, given a station ID}
\usage{
station_today(station_id, timezone = -1, dir_numeric = FALSE)
}
\arguments{
\item{station_id}{The station identifier (ex: KDEN, KNYC, etc).}

\item{timezone}{The nominal timezone for the forecast. One of \code{OlsonNames()} or \code{-1} for local time. Defaults to \code{-1}.}

\item{dir_numeric}{\code{TRUE} for numeric directions, \code{FALSE} for character directions; defaults to \code{FALSE}.}
}
\value{
A sf object with timestamps and weather values
}
\description{
Get todays weather data, given a station ID
}
\examples{
station_today("KNYC")
}
