% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-station_forecast.R
\name{station_tomorrow}
\alias{station_tomorrow}
\title{Get tomorrows forecast for a given station_id}
\usage{
station_tomorrow(station_id, timezone = -1, dir_numeric = FALSE, short = TRUE)
}
\arguments{
\item{station_id}{An ASOS or AWOS station identifier}

\item{timezone}{A timezone to denominate the forecast in}

\item{dir_numeric}{logical; if TRUE return the wind direction as numeric. If FALSE return the wind direction as a character}

\item{short}{logical; if TRUE return the forecast for only tomorrow. If FALSE return the forecast for the remainder of today and tomorrow.}
}
\value{
Returns a sf with coordinates for the AWOS/ASOS station, time, and forecast values
}
\description{
Get tomorrows forecast for a given station_id
}
\examples{
station_tomorrow("KBOS")

}
