% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-station_data.R
\name{station_tz}
\alias{station_tz}
\title{Fetch station timezone}
\usage{
station_tz(station_id)
}
\arguments{
\item{station_id}{The station identifier (ex: KDEN, KBOS, KNYC, etc).}
}
\value{
A character corresponding to a timezone from \code{OlsonNames()}.
}
\description{
Fetch station timezone
}
\examples{
\donttest{
Sys.time() \%>\% lubridate::force_tz(tzone = station_tz("KDEN"))
}
}
